+++ Assets/Scripts/Core/Input/ControlScheme.cs
@@ -2,7 +2,10 @@
 {
     public enum ControlScheme
     {
+        Unknown = -1,
         KeyboardMouse = 0,
         Gamepad = 1,
+        Touch = 2,
+        Joystick = 3,
     }
 }