+++ Assets/Scripts/Runtime/Stats/ItemUsedStat.cs
@@ -1,0 +1,14 @@
+﻿using RIEVES.GGJ2026.Runtime.Items;
+
+namespace RIEVES.GGJ2026.Runtime.Stats
+{
+    internal readonly struct ItemUsedStat
+    {
+        public ItemData ItemData { get; }
+
+        public ItemUsedStat(ItemData itemData)
+        {
+            ItemData = itemData;
+        }
+    }
+}